+++ frontend/src/app/main/ui/workspace/main_menu.cljs
@@ -618,7 +618,7 @@
         [:> dropdown-menu-item* {:class (stl/css :submenu-item)
                                  :on-click    on-pin-version
                                  :on-key-down on-pin-version-key-down
-                                 :id          "file-menu-show-version-history"}
+                                 :id          "file-menu-create-version"}
          [:span {:class (stl/css :item-name)}
           (tr "dashboard.create-version-menu")]]
 
@@ -627,7 +627,10 @@
                                  :on-key-down on-show-version-history-key-down
                                  :id          "file-menu-show-version-history"}
          [:span {:class (stl/css :item-name)}
-          (tr "dashboard.show-version-history")]]
+          (tr "dashboard.show-version-history")]
+         [:span {:class (stl/css :shortcut)}
+          (for [sc (scd/split-sc (sc/get-tooltip :toggle-history))]
+            [:span {:class (stl/css :shortcut-key) :key sc} sc])]]
 
         [:div {:class (stl/css :separator)}]])